requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asterisk-ami==0.1.6
asyncpg==0.28.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
gevent==23.9.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-multipart==0.0.6
requests==2.31.0
sentry-sdk==1.34.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
websocket-client==1.6.4
websockets==12.0
zope.event==5.0
zope.interface==6.1
